+++ src/if_cscope.c
@@ -540,7 +540,7 @@ cs_add_common(
     char	*fname2 = NULL;
     char	*ppath = NULL;
     int		i;
-    int		len;
+    size_t	len;
     size_t	usedlen = 0;
     char_u	*fbuf = NULL;
 
@@ -549,7 +549,7 @@ cs_add_common(
 	goto add_err;
 
     expand_env((char_u *)arg1, (char_u *)fname, MAXPATHL);
-    len = (int)STRLEN(fname);
+    len = STRLEN(fname);
     fbuf = (char_u *)fname;
     (void)modify_fname((char_u *)":p", FALSE, &usedlen,
 					      (char_u **)&fname, &fbuf, &len);
